% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_form_data.R
\name{build_form_data}
\alias{build_form_data}
\title{Construir Dados de Formulário para Requisição}
\usage{
build_form_data(args_processados, base_form)
}
\arguments{
\item{args_processados}{Dataframe contendo argumentos processados.}

\item{base_form}{Objeto contendo as informações da URL do formulário.}
}
\value{
Uma string formatada pronta para ser usada em uma requisição web.
}
\description{
Esta função constrói uma string de consulta para um formulário web a partir de
argumentos processados e informações de uma base de dados. Ela ajusta os caracteres
especiais para o formato apropriado para URLs e cria o link de requisição final.
}
\examples{
\dontrun{
# Exemplo em construção...
# args_processados <- data.frame(argument = c("arg1", "arg2"), value = c("val1", "val2"))
# base <- list(url_form = "http://example.com/form")
# build_form_data(args_processados, base)

}
}
