% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catoper-package.R
\docType{package}
\name{catoper-package}
\alias{catoper}
\alias{catoper-package}
\title{catoper: Testes de funções}
\description{
The catoper is a miscellany of functions useful to myself.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/denis-or/catoper}
  \item Report bugs at \url{https://github.com/denis-or/catoper/issues}
}

}
\author{
\strong{Maintainer}: Denis Rodrigues \email{catoper@gmail.com} (\href{https://orcid.org/0000-0003-3748-876X}{ORCID})

}
\keyword{internal}
