% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_.R
\name{check_}
\alias{check_}
\title{Verificar Presença de Argumento e Valor em um Dataframe}
\usage{
check_(argument, value_to_check, base_toda_z)
}
\arguments{
\item{argument}{O nome do argumento a ser verificado no dataframe.}

\item{value_to_check}{O valor do argumento a ser verificado.}

\item{base_toda_z}{Dataframe que contém os dados a serem verificados.}
}
\value{
Um dataframe contendo as linhas onde o argumento e seu valor correspondente
são encontrados em \code{base_toda_z}.
}
\description{
Esta função verifica se um dado argumento e seu valor correspondente estão presentes
em um dataframe \code{base_toda}. Se estiverem presentes, retorna um subset do dataframe.
Caso contrário, interrompe a execução com uma mensagem de erro.
}
\examples{
\dontrun{
# Exemplo em construção...
# check_("argumento_exemplo", "valor_exemplo", base_toda)

}
}
