% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_base_toda.R
\name{create_base_toda}
\alias{create_base_toda}
\title{Recriar Formulário a partir de Página HTML}
\usage{
create_base_toda(page_x)
}
\arguments{
\item{page_x}{Objeto de página HTML de onde as informações serão extraídas.}
}
\value{
Um dataframe contendo informações dos campos de formulário, incluindo
campos, valores e etiquetas correspondentes.
}
\description{
Esta função raspa informações de uma página HTML e recria um formulário de dados.
Extrai informações dos elementos 'select' e 'option', organiza esses dados em um
dataframe e adiciona etiquetas correspondentes a cada campo.
}
\examples{
\dontrun{
# Exemplo em construção...

}
}
