% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data_datasus.R
\name{fetch_data_datasus}
\alias{fetch_data_datasus}
\title{Baixar e Processar Dados}
\usage{
fetch_data_datasus(form_data, base, linha_argumento)
}
\arguments{
\item{form_data}{Dados de formulário pré-formatados para a requisição POST.}

\item{base}{Lista contendo informações da URL base e do formulário.}

\item{linha_argumento}{Especifica a linha de início para processamento dos dados no CSV.}
}
\value{
Um dataframe contendo os dados processados do CSV.
}
\description{
Esta função é responsável por fazer uma requisição POST ao site do Tabnet escolhido,
baixar um arquivo CSV e processá-lo. Ela tenta a requisição várias vezes
em caso de timeout e processa o CSV baixado para retornar os dados em um formato
utilizável.
}
\examples{
\dontrun{
# Exemplo em construção...

}
}
