% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_char}
\alias{replace_char}
\title{Substituir Caracteres Especiais em uma String}
\usage{
replace_char(string)
}
\arguments{
\item{string}{Uma string que pode conter caracteres especiais.}
}
\value{
Uma string com os caracteres especiais substituídos por
seus códigos percentuais.
}
\description{
Esta função substitui caracteres especiais em uma string pelos seus
respectivos códigos percentuais. É útil para preparar strings para
serem usadas em URLs ou outros contextos que exigem codificação de caracteres.
}
\examples{
\dontrun{
replace_char("Olá, Mundo!")
replace_char("Exemplo com # e @.")
}
}
