% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_progress}
\alias{show_progress}
\title{Exibir uma Barra de Progresso}
\usage{
show_progress(msg, time_div)
}
\arguments{
\item{msg}{Uma mensagem para exibir junto com a barra de progresso.}

\item{time_div}{A duração total durante a qual a barra de progresso será exibida.
Este valor é dividido igualmente entre os 100 passos da barra.}
}
\description{
Esta função exibe uma barra de progresso no console. É útil para indicar o progresso
de uma tarefa que leva um certo tempo para ser concluída.
}
\examples{
\dontrun{
# Exibe uma barra de progresso por 5 segundos
show_progress("Processando", 5)
}
}
