% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_base_readme.R
\name{update_base_readme}
\alias{update_base_readme}
\title{Atualizar base usada no \code{Readme}}
\usage{
update_base_readme()
}
\value{
Um dataframe contendo informações dos eventos alcançados pelo pacote
e suas atualizações.
}
\description{
Esta função raspa informações das páginas HTML e gera um compilado de dados
para acompanhamento do usuário, apresentado em forma de tabela no \code{Readme}.
}
