% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data_datasus.R
\name{fetch_data_datasus}
\alias{fetch_data_datasus}
\title{Baixar e Processar Dados do Datasus}
\usage{
fetch_data_datasus(form_data, base, linha_argumento)
}
\arguments{
\item{form_data}{Dados de formulário pré-formatados para a requisição POST.}

\item{base}{Lista contendo informações da URL base e do formulário.}

\item{linha_argumento}{Especifica a linha de início para processamento dos dados no CSV.}
}
\value{
Um dataframe contendo os dados processados do CSV.
}
\description{
Esta função é responsável por fazer uma requisição POST ao Datasus,
baixar um arquivo CSV e processá-lo. Ela tenta a requisição várias vezes
em caso de timeout e processa o CSV baixado para retornar os dados em um formato
utilizável.
}
\examples{
\dontrun{
# Exemplo em construção...
# form_data <- "seus_dados_de_formulario"
# base_exemplo <- list(url_tbcgi = "http://exemplo.com", url_form = "http://exemplo.com/form")
# linha_argumento <- "linha_inicial"
# fetch_data_datasus(form_data, base_exemplo, linha_argumento)

}
}
