% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event.R
\name{get_event}
\alias{get_event}
\title{Gerar formulário necessário para conexão com o Tabnet do Datasus}
\usage{
get_event(event)
}
\arguments{
\item{event}{Nome do evento para o qual os dados devem ser baixados e processados.
Deve ser uma string não vazia.}
}
\value{
Uma lista contendo a página HTML lida e os metadados filtrados para o evento.
}
\description{
Esta função é usada para leitura do formulário para tabulação de dados
de evento de saúde do Tabnet, no Datasus. Ela realiza o download do html
do formulário do Tabnet para que os argumentos sejam validados.
A lista de eventos é padronizada e pode ser encontrada \href{event_list.html}{neste link}.
}
\details{
IMPORTANT:
Essa função é essencial para uso da função \code{pega_dados()}.
}
\examples{
\dontrun{
# Exemplo em construção...
dados_evento <- get_event("Influenza")

}
}
\author{
Denis Rodrigues \email{catoper@gmail.com}
}
