% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event.R
\name{get_event}
\alias{get_event}
\title{Gerar formulário necessário para conexão com o Tabnet do Datasus}
\usage{
get_event(event)
}
\arguments{
\item{event}{Nome do evento para o qual os dados devem ser baixados e processados.
Deve ser uma string não vazia.}
}
\value{
Uma lista contendo a página HTML lida e os metadados filtrados para o evento.
}
\description{
Esta função é usada para leitura do formulário para tabulação de dados
de evento de saúde do Tabnet, no Datasus. Ela realiza o download do html
do formulário do Tabnet para que os argumentos sejam validados.
A lista de eventos é padronizada e pode ser encontrada \href{event_list.html}{neste link}.
}
\details{
IMPORTANT:
Essa função é essencial para uso da função \code{pega_dados()}.
}
\examples{
\dontrun{

# Obter dados para um evento específico
# Primeiro, obter a página para o evento desejado
pagina <- get_event("Imunização Doses Aplicadas")

# Em seguida, usar pega_dados para obter a tabela com os parâmetros específicos
# Observação: os argumentos devem ser idênticos àqueles que estão disponíveis
# no Tabnet do Datasus. Apenas o argumento `ob` é obrigatório e constante.

tabela_D1 <- pega_dados(
  ob = pagina,
  Linha = "Município",
  Coluna = "Dose",
  Medidas = "Doses_aplicadas",
  "Períodos Disponíveis" = 2022,
  "Dose" = "1ª dose",
  "Imunobiológicos" = "Poliomielite inativada (VIP)",
  "Unidade da Federação" = "Rio de Janeiro"
)
}

# A tabela_D1 agora contém os dados processados de doses aplicadas de
# 1ª dose de polio (VIP) em 2022 nos municípios do Rio de Janeiro




}
\author{
Denis Rodrigues \email{catoper@gmail.com}
}
