% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/see_normality.R
\name{see_normality}
\alias{see_normality}
\title{Check the normality of the data}
\usage{
see_normality(variable, label)
}
\arguments{
\item{variable}{vector or variable with the data to check}

\item{label}{label to plots}
}
\value{
a beautiful plot
}
\description{
Check the normality of the data
}
